Use alternative github action
I prefer this to other actions at the code is simple and it uses a
deploy key that is limited to this repository unlike all the other
variants which require a person token with wide repository access.

See the following for information on how this was created:
https://github.com/khanhicetea/gh-actions-hugo-deploy-gh-pages/blob/7ce03ccc8ea878ecf739e4f046243d2855a670be/README.md#create-deploy-key

It also means this key isn't bound to an employee (me) which wouldn't be
very considerate.