Fetch post dates from git
Instead of relying on manually updating the date on each post I'm
changing it to update based on the last change date for that file in
Git.

The reason is that people generally forget these things and we want to
keep these posts as living documents so the date should be changing
pretty often so it should be simple.