Replace hugo-primer until it's fixed upstream
I've opened a merge request for a fix to a bug in the theme we are using
which causes Hugo to crash when you try to add a child menu entry.

Until it's merged I'm switching to my fork that has the fix.

Here is the failing build which shows the error:
https://app.netlify.com/sites/runbook-builder/deploys/5ecfefd20903b900075c5592

The theme creator is trying to access a variable that doesn't exist in
the menu context.