This is a test to see if this makes data unidentifiable
I found this information in an old Hugo ticket that mentions a no-cookie
version of the Google Analytics template and how to enable it. I'm
trying this to see what it produces in the code.

This seems to force GA to use SessionStorage which expires as soon a the
tab is closed:
https://developer.mozilla.org/en-US/docs/Web/API/Window/sessionStorage

I think this is a better situation in terms of user privacy, but I still
think we probably need a cookie consent before using Google Analytics.
I'm just not sure how we can implement this within Hugo. It may need
changes to Hugo itself.