Update hugo-primer theme to add header-start
This update is required so we can inject the Osano script at the
beginning of the HTML head tag. This is required so that it can block
other scripts from setting cookies before we have permission from the
Vistor.

I'm also adding the required script tag so that Osano can be loaded as
this shows an example of what the new hugo-primer version allows.