Remove Disqus comments from the policy page
This fixes https://github.com/ContainerSolutions/runbooks/issues/50
We are lucky and dispite Hugo not having anything official there is a
currently undocumented flag in the theme that we are using.

See here for more info:
https://github.com/gohugoio/hugo/issues/7373#issuecomment-642535593

Anyway, we don't want people leaving comments on the privacy policy as
it just doesn't seem like a good idea to mix user content with a legal
document.