Fix bug with hugo not respecting baseurl in static files
This is a temporary work around for what I suspect is a bug in the theme
we are using. For some reason image paths are not being prefixed by the
necessary basepath "/runbooks" or "/" on Netlify.

This shortcode works around this issue but I don't remember having this
problem with the previous theme, so I suspect this is where the real
issue lies and needs to be fixed. However for now I'm just going to use
this shortcode instead.