Account for the current WorkDir in sys.path
The initial directions would lead to a blank first entry which may be
where the problem lies if the user is trying to load a local module but
running their script using an absolute path from a different directory.

I also removed the ">>>>" at the beginning of the command as we are now
suggesting using it in a file, but also because it makes copy and paste
of the lines much easier.

Lastly I'm splitting each paths onto seperate lines for easier reading.