Add "Improve this page" link
For some _index pages this isn't appropriate as it normally only
displays the title with the body being autogenerated. For such cases I'm
asking the user to "Suggest a change" instead and linking them to the
issues page.

I've included a free icon for the pencil from Github's Octicons
(https://primer.style/octicons/) resource. It uses the MIT license
and we are not using "substantial portions of the Software". However
here is the link to that license anyway:
https://github.com/primer/octicons/blob/e9a9a84fb796d70c0803ab8d62eda5c03415e015/LICENSE

I've inlined the SVG as it's currently only being used in one location
and is quite small so I believe this leads to faster loading.