Add a fingerprint to the CSS override
When recently making an update to the stylesheet I saw that the changes
were not immediately updated in my browser. Most frameworks add
automatic fingerprints to the assets they include so that if there is a
change to the file contents, then filename also changes. This invalidates
any cache of the previous version.

Hugo unfortunately doesn't automatically do this. This change introduces
the method recommended by Hugo to address this issue.

See: https://gohugo.io/hugo-pipes/fingerprint/

I'm using md5 as it's faster and shorter while still perfect for cache
invalidation. The Hugo docs are also using the hash for security reasons
and so they went with sha-512, which makes sense if that is your use case.

See the following issue where a gentleman named Fjan painstakingly
explains the difference:
https://github.com/rails/sprockets/issues/124