Fix warning about .File not existing
I've noticed a warning has started showning up:
`.File.Filename on zero object. Wrap it in if or with: {{ with .File }}{{ .Filename }}{{ end }}`

Hugo unfortunately doesn't tell me what page is leading to this error,
but either way the following code should fix things and the logic still
works when I tested it.