Add theme to help with debugging Hugo
There has unfortunately been some pushback with adding general debugging
tools to Hugo

https://github.com/gohugoio/hugo/issues/4081#issuecomment-442384273
https://github.com/gohugoio/hugo/issues/3957#issuecomment-364657015

I've found this can make things a lot more difficult than it needs
to be. There appears to be a belief that adding debugging tools will
lead to people using things that are undocumented. Unfortunately in my
experience I've had trouble with values that are documented but where
the documentation isn't very clear.

Even where debugging tools are provided the output is minified and
requests to change this have so far been rejected for the same reason as
above.

A user has developed a theme which can work around many of these issues
and add a convenient debugging partial to allow you to clearly see the
contents of most variables. (see README changes for more specifics.)